+++ packages/zowe-explorer/src/dataset/actions.ts
@@ -158,7 +158,6 @@ export async function allocateLike(datasetProvider: api.IZoweTree<api.IZoweDatas
 
     const theFilter = datasetProvider.createFilterString(newDSName, currSession);
     currSession.tooltip = currSession.pattern = theFilter.toUpperCase();
-    datasetProvider.addSearchHistory(theFilter);
     datasetProvider.refresh();
     currSession.dirty = true;
     datasetProvider.refreshElement(currSession);